+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/item/PylonItemSchema.kt
@@ -1,7 +1,7 @@
 package io.github.pylonmc.pylon.core.item
 
 import io.github.pylonmc.pylon.core.item.PylonItem.Companion.idKey
-import io.github.pylonmc.pylon.core.persistence.PylonSerializers
+import io.github.pylonmc.pylon.core.persistence.datatypes.PylonSerializers
 import io.github.pylonmc.pylon.core.registry.PylonRegistry
 import org.bukkit.Keyed
 import org.bukkit.NamespacedKey
