+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/persistence/datatypes/DelegatingPersistentDataType.kt
@@ -1,0 +1,27 @@
+package io.github.pylonmc.pylon.core.persistence.datatypes
+
+import org.bukkit.persistence.PersistentDataAdapterContext
+import org.bukkit.persistence.PersistentDataType
+
+/**
+ * A [PersistentDataType] that first converts the complex type
+ * to another type, which then uses the delegate to convert to the primitive type.
+ */
+abstract class DelegatingPersistentDataType<P : Any, D : Any, C : Any>(
+    val delegate: PersistentDataType<P, D>,
+) : PersistentDataType<P, C> {
+
+    final override fun getPrimitiveType(): Class<P> = delegate.primitiveType
+
+    abstract fun toDelegatedType(complex: C): D
+
+    abstract fun fromDelegatedType(primitive: D): C
+
+    final override fun toPrimitive(complex: C, context: PersistentDataAdapterContext): P {
+        return delegate.toPrimitive(toDelegatedType(complex), context)
+    }
+
+    final override fun fromPrimitive(primitive: P, context: PersistentDataAdapterContext): C {
+        return fromDelegatedType(delegate.fromPrimitive(primitive, context))
+    }
+}
