+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/persistence/datatypes/PylonSerializers.kt
@@ -1,0 +1,112 @@
+package io.github.pylonmc.pylon.core.persistence.datatypes
+
+import io.github.pylonmc.pylon.core.block.BlockPosition
+import io.github.pylonmc.pylon.core.block.ChunkPosition
+import io.github.pylonmc.pylon.core.persistence.datatypes.BlockPositionPersistentDataType
+import io.github.pylonmc.pylon.core.persistence.datatypes.CharPersistentDataType
+import io.github.pylonmc.pylon.core.persistence.datatypes.ChunkPositionPersistentDataType
+import io.github.pylonmc.pylon.core.persistence.datatypes.LocationPersistentDataType
+import io.github.pylonmc.pylon.core.persistence.datatypes.NamespacedKeyPersistentDataType
+import io.github.pylonmc.pylon.core.persistence.datatypes.UUIDPersistentDataType
+import io.github.pylonmc.pylon.core.persistence.datatypes.VectorPersistentDataType
+import io.github.pylonmc.pylon.core.persistence.datatypes.WorldPersistentDataType
+import io.github.pylonmc.pylon.core.pluginInstance
+import org.bukkit.Bukkit
+import org.bukkit.Location
+import org.bukkit.NamespacedKey
+import org.bukkit.World
+import org.bukkit.persistence.PersistentDataAdapterContext
+import org.bukkit.persistence.PersistentDataContainer
+import org.bukkit.persistence.PersistentDataType
+import org.bukkit.util.Vector
+import java.nio.ByteBuffer
+import java.util.*
+
+object PylonSerializers {
+    @JvmField
+    val BYTE = PersistentDataType.BYTE!!
+
+    @JvmField
+    val SHORT = PersistentDataType.SHORT!!
+
+    @JvmField
+    val INTEGER = PersistentDataType.INTEGER!!
+
+    @JvmField
+    val LONG = PersistentDataType.LONG!!
+
+    @JvmField
+    val FLOAT = PersistentDataType.FLOAT!!
+
+    @JvmField
+    val DOUBLE = PersistentDataType.DOUBLE!!
+
+    @JvmField
+    val BOOLEAN = PersistentDataType.BOOLEAN!!
+
+    @JvmField
+    val STRING = PersistentDataType.STRING!!
+
+    @JvmField
+    val CHAR = CharPersistentDataType
+
+    @JvmField
+    val BYTE_ARRAY = PersistentDataType.BYTE_ARRAY!!
+
+    @JvmField
+    val INTEGER_ARRAY = PersistentDataType.INTEGER_ARRAY!!
+
+    @JvmField
+    val LONG_ARRAY = PersistentDataType.LONG_ARRAY!!
+
+    @JvmField
+    val TAG_CONTAINER = PersistentDataType.TAG_CONTAINER!!
+
+    @JvmField
+    val NAMESPACED_KEY = NamespacedKeyPersistentDataType
+
+    @JvmField
+    val UUID = UUIDPersistentDataType
+
+    @JvmField
+    val VECTOR = VectorPersistentDataType
+
+    @JvmField
+    val WORLD = WorldPersistentDataType
+
+    @JvmField
+    val BLOCK_POSITION = BlockPositionPersistentDataType
+
+    @JvmField
+    val CHUNK_POSITION = ChunkPositionPersistentDataType
+
+    @JvmField
+    val LOCATION = LocationPersistentDataType
+
+    @JvmField
+    val LIST = PersistentDataType.LIST!!
+
+    @JvmField
+    val SET = SetPersistentDataType.Companion
+
+    @JvmField
+    val MAP = MapPersistentDataType.Companion
+}
+
+
+object CharPersistentDataType : PersistentDataType<ByteArray, Char> {
+    override fun getPrimitiveType(): Class<ByteArray> = ByteArray::class.java
+
+    override fun getComplexType(): Class<Char> = Char::class.java
+
+    override fun fromPrimitive(primitive: ByteArray, context: PersistentDataAdapterContext): Char {
+        val buffer = ByteBuffer.wrap(primitive)
+        return buffer.getChar()
+    }
+
+    override fun toPrimitive(complex: Char, context: PersistentDataAdapterContext): ByteArray {
+        val buffer = ByteBuffer.allocate(Char.SIZE_BYTES)
+        buffer.putChar(complex)
+        return buffer.array()
+    }
+}
