+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/generictest/PylonPDCPrimitivesTest.java
@@ -1,7 +1,7 @@
 package io.github.pylonmc.pylon.test.generictest;
 
 import io.github.pylonmc.pylon.core.persistence.PylonPersistentDataContainer;
-import io.github.pylonmc.pylon.core.persistence.PylonSerializers;
+import io.github.pylonmc.pylon.core.persistence.datatypes.PylonSerializers;
 import io.github.pylonmc.pylon.test.GenericTest;
 import io.github.pylonmc.pylon.test.TestAddon;
 import org.bukkit.NamespacedKey;
